requirements.txt
flask
python-pptx
pillow
requests
